+++ src/dockerinside/setup/__init__.py
@@ -66,6 +66,10 @@ class SetupApp(dockerutils.BasicDockerApp):
                             help="Pull unavailable images automatically")
         parser.add_argument('--refspec',
                             help="Refspec for su-exec repo (tag/branch; default: master)")
+        parser.add_argument('--host-network',
+                            action="store_true",
+                            default=False,
+                            help="Allow access to host network (f.e. if using a proxy on locahost)")
         parser.set_defaults(loglevel=logging.INFO)
         args = parser.parse_args(args=argv)
         return args
@@ -108,12 +112,15 @@ class SetupApp(dockerutils.BasicDockerApp):
         host_env = dict({k:v for k,v in os.environ.items() if k in self.PASSED_HOST_ENV})
         env.update(host_env)
         logging.debug("Prepared environment: %s", host_env)
+        network_mode = 'host' if self._args.host_network else None
+        logging.debug("Network mode: %s", "default" if network_mode is None else network_mode)
         cobj = self._dc.containers.create(
             self.DEFAULT_IMAGE,
             command="/entrypoint.sh",
             volumes=volumes,
             environment=env,
             name=name,
+            network=network_mode
         )
         try:
             cobj.put_archive('/', script_pack)
@@ -121,27 +128,34 @@ class SetupApp(dockerutils.BasicDockerApp):
             for msg in cobj.logs(stdout=True, stderr=True, stream=True):
                 logging.debug("{0}".format(msg.decode('utf-8').rstrip('\n')))
             ret = cobj.wait()
-            logging.info("setup returned {0}".format(ret['StatusCode']))
+            status_code = ret.get('StatusCode', None)
+            logging.info("setup returned %s", status_code)
+            return status_code
         finally:
             cobj.stop()
             cobj.remove()
 
     def run(self, argv):
+        ret = 1
         self._args = self._parse_args(argv)
         logging.getLogger().setLevel(self._args.loglevel)
         # noinspection PyBroadException
         try:
-            self.setup(self._args.url,
-                       home=self._args.home,
-                       auto_pull=self._args.auto_pull,
-                       name=self._args.name,
-                       refspec=self._args.refspec)
+            ret = self.setup(
+                self._args.url,
+                home=self._args.home,
+                auto_pull=self._args.auto_pull,
+                name=self._args.name,
+                refspec=self._args.refspec
+            )
         except dockerutils.InvalidPath as e:
             logging.exception("{0} '{1}' doesn't exist".format(e.type_, e.path))
         except docker.errors.ImageNotFound:
             logging.exception("Image '{0}' not found".format(self.DEFAULT_IMAGE))
         except Exception:
             logging.exception("Failed to run setup()")
+        finally:
+            return ret
 
 
 def setup_main():